Add Integer and text representations

+++ composeApp/src/commonMain/kotlin/fr/nicopico/petitboutiste/models/RepresentationFormat.kt
@@ -5,7 +5,7 @@ import java.nio.charset.Charset
 sealed class RepresentationFormat {
     data object Hexadecimal : RepresentationFormat()
 
-    data class Decimal(
+    data class Integer(
         val endianness: Endianness = Endianness.BigEndian,
     ) : RepresentationFormat()
